Separate df_module from argument parsing.
